feat:🎸 move shared component in lib
