+++ src/app/components/stepper-steps/stepper-steps.component.stories.ts
@@ -3,7 +3,7 @@ import { StepperStepsComponent } from './stepper-steps.component';
 import { AsyncPipe, CommonModule } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { RouterLink } from '@angular/router';
-import { ButtonComponent } from '../../shared/button/button.component';
+import { ButtonComponent } from '../../../../ftx-forms/src/lib/ftx-forms/shared/button/button.component';
 import { MatRippleModule } from '@angular/material/core';
 import { Store } from '@ngrx/store';
 import { StepperService } from '../services/stepper.service';
